requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
selenium==4.19.0
python-multipart==0.0.9
openpyxl==3.1.2
pydantic==2.6.4
pydantic-settings==2.2.1
bcrypt==4.1.2
cryptography==42.0.5
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0